Make DropSourceBehavior partial and split BehaviorDataObject into its own file
